pyproject.toml
[tool.poetry]
name = "ui-automation-training"
version = "0.1.0"
description = ""
authors = ["Wojciech Hubczak <wojciech.hubczak.external@eviden.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.7.0"
pytest-bdd = "^8.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = 6.0
addopts = "-ra -q --cucumber-json=test.json --tracing=retain-on-failure"
testpaths = "tests/"
python_files = "*step_definitions.py"